+++ edd-sl-updater.php
@@ -11,14 +11,14 @@
  * Plugin Name:       EDD Software Licensing Updater
  * Plugin URI:        https://github.com/afragen/edd-sl-updater
  * Description:       A universal updater for EDD Software Licensing products.
- * Version:           0.11.9
+ * Version:           0.11.10
  * Author:            Andy Fragen
  * License:           MIT
  * Network:           true
  * Text Domain:       edd-sl-updater
  * Domain Path:       /languages
  * GitHub Plugin URI: https://github.com/afragen/edd-sl-updater
- * Requires at least: 4.6
+ * Requires at least: 4.7
  * Requires PHP:      5.6
  */